requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.4.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.9
dataclasses-json==0.6.4
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
easydict==1.13
fairscale==0.4.13
filelock==3.12.4
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.9.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.4
imageio==2.34.0
iniconfig==2.0.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
llama-index-core==0.10.18.post1
llama-index-embeddings-huggingface==0.1.4
llama-index-llms-ollama==0.1.2
llama-index-readers-file==0.1.8
llamaindex-py-client==0.1.13
MarkupSafe==2.1.3
marshmallow==3.21.1
matplotlib==3.8.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
mypy-extensions==1.0.0
neovim==0.3.1
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
nose==1.3.7
numpy==1.26.0
onnx==1.15.0
openai==1.13.3
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.1
Pillow==10.0.1
pluggy==1.3.0
prettytable==3.10.0
protobuf==5.26.0
pydantic==2.6.3
pydantic_core==2.16.3
pyflakes==3.1.0
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pynvim==0.5.0
pyparsing==3.1.2
pypdf==4.1.0
PyPDF2==3.0.1
pytest==7.4.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tifffile==2024.2.12
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.0.dev20230917
torchaudio==2.2.0.dev20230917
torchvision==0.17.0.dev20230917
tqdm==4.66.2
transformers==4.38.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.4
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
